+++ blocks/docs-search/index.asset.php
@@ -1,0 +1,11 @@
+<?php
+return array(
+	'dependencies' => array(
+		'wp-api-fetch',
+		'wp-block-editor',
+		'wp-blocks',
+		'wp-components',
+		'wp-element',
+	),
+	'version'      => DOCSRAPTOR_VERSION,
+);